requirements.txt
Pillow
Flask
Flask-Cache
Flask-SQLAlchemy
requests_oauthlib
